+++ Anlab.Core/Models/OrderDetails.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Anlab.Core.Domain;
+using Anlab.Core.Models;
+
+
+namespace AnlabMvc.Models.Order
+{
+    public class OrderDetails
+    {
+        public OrderDetails()
+        {
+
+        }
+
+        public int Quantity { get; set; }
+        public string SampleType { get; set; }
+        public string AdditionalInfo { get; set; }
+        public IList<TestItem> SelectedTests { get; set; }
+
+        public Decimal Total { get; set; }
+
+        public Payment Payment { get; set; }
+    }
+
+}
