+++ Anlab.Core/Models/Payment.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Anlab.Core.Models
+{
+    public class Payment
+    {
+        public string ClientType { get; set; }
+        public string Account { get; set; }
+    }
+}
